fix: resolve CRITICAL data races, SPSC queue bug, and hot-path heap allocations
- SPSCQueue: monotonic counter algorithm (fix silent data corruption)
- StorageRouter: channel_id-keyed maps (zero heap alloc on hot path)
- StorageRouter: mutex-guarded CSV/MMAP paths (fix data race)
- StorageRouter: DolphinDB I/O moved outside buffer mutex
- PubWorker: acquire/release ordering on running_ flag
- PubWorker: remove redundant TickData copy+memset on hot path
- PubWorker: adaptive idle wait (yield→sleep to reduce latency)
- CsvWriter: fmt::memory_buffer replaces std::to_string (zero alloc)
- CsvWriter: EnsureDirExists checks mkdir return value
- SecureString: password zeroed on destruction
- Fix default config path, stale comments, include paths, BuildTopic API
- Add tests: SPSCQueue (5), StorageRouter (5), CsvWriter (4)

+++ src/pubsub/pub_message.h
@@ -35,11 +35,11 @@ constexpr size_t kMaxTopicLen = 64;
 // Example:      "binance:spot:BTCUSDT:tick"
 //
 // Returns the number of bytes written (excluding null terminator).
-// Returns 0 if the buffer is too small (truncation would occur).
-size_t BuildTopic(char* buf, size_t buf_size,
-                  std::string_view exchange,
-                  std::string_view channel_type,
-                  std::string_view symbol,
-                  std::string_view event_type);
+// Returns -1 if the buffer is too small (truncation would occur).
+int64_t BuildTopic(char* buf, size_t buf_size,
+                   std::string_view exchange,
+                   std::string_view channel_type,
+                   std::string_view symbol,
+                   std::string_view event_type);
 
 }  // namespace sqc
